feat: revise test code for auth user
