+++ src/test/java/com/scg/stop/controller/EventNoticeControllerTest.java
@@ -1,13 +1,14 @@
-package com.scg.stop.domain.notice.controller;
+package com.scg.stop.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scg.stop.configuration.AbstractControllerTest;
 import com.scg.stop.domain.file.dto.response.FileResponse;
-import com.scg.stop.notice.controller.NoticeController;
-import com.scg.stop.notice.dto.request.NoticeRequest;
-import com.scg.stop.notice.dto.response.NoticeListElementResponse;
-import com.scg.stop.notice.dto.response.NoticeResponse;
-import com.scg.stop.notice.service.NoticeService;
+import com.scg.stop.event.controller.EventNoticeController;
+import com.scg.stop.event.dto.request.EventNoticeRequest;
+import com.scg.stop.event.dto.response.EventNoticeListElementResponse;
+import com.scg.stop.event.dto.response.EventNoticeResponse;
+import com.scg.stop.event.service.EventNoticeService;
+import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +18,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
+import org.springframework.http.HttpHeaders;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.ResultActions;
 
@@ -29,63 +30,79 @@ import java.util.List;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
+import static org.springframework.restdocs.cookies.CookieDocumentation.cookieWithName;
+import static org.springframework.restdocs.cookies.CookieDocumentation.requestCookies;
+import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(NoticeController.class)
+@WebMvcTest(EventNoticeController.class)
 @MockBean(JpaMetamodelMappingContext.class)
 @AutoConfigureRestDocs
-class NoticeControllerTest extends AbstractControllerTest {
+class EventNoticeControllerTest extends AbstractControllerTest {
 
+    private static final String ACCESS_TOKEN = "admin_access_token";
+    private static final String REFRESH_TOKEN = "refresh_token";
     @MockBean
-    private NoticeService noticeService;
+    private EventNoticeService eventNoticeService;
 
     @Autowired
     private ObjectMapper objectMapper;
 
     @Test
-    @DisplayName("공지 사항을 생성할 수 있다.")
+    @DisplayName("이벤트 공지 사항을 생성할 수 있다.")
         // TODO: Auth user check
-    void createNotice() throws Exception {
+    void createEventNotice() throws Exception {
 
         // given
-        NoticeRequest request = new NoticeRequest("공지 사항 제목", "공지 사항 내용", true, List.of(1L, 2L, 3L));
+        EventNoticeRequest request = new EventNoticeRequest("이벤트 공지 사항 제목", "이벤트 공지 사항 내용", true, List.of(1L, 2L, 3L));
 
         List<FileResponse> files = Arrays.asList(
                 new FileResponse(1L, "014eb8a0-d4a6-11ee-adac-117d766aca1d", "예시 첨부 파일 1.jpg", "image/jpeg", LocalDateTime.now(), LocalDateTime.now()),
                 new FileResponse(2L, "11a480c0-13fa-11ef-9047-570191b390ea", "예시 첨부 파일 2.jpg", "image/jpeg", LocalDateTime.now(), LocalDateTime.now()),
                 new FileResponse(3L, "1883fc70-cfb4-11ee-a387-e754bd392d45", "예시 첨부 파일 3.jpg", "image/jpeg", LocalDateTime.now(), LocalDateTime.now())
         );
-        NoticeResponse response = new NoticeResponse(1L, "공지 사항 제목", "공지 사항 내용", 0, true, LocalDateTime.now(), LocalDateTime.now(), files);
+        EventNoticeResponse response = new EventNoticeResponse(1L, "이벤트 공지 사항 제목", "이벤트 공지 사항 내용", 0, true, LocalDateTime.now(), LocalDateTime.now(), files);
 
-        when(noticeService.createNotice(any(NoticeRequest.class))).thenReturn(response);
+        when(eventNoticeService.createEventNotice(any(EventNoticeRequest.class))).thenReturn(response);
 
         // when
         ResultActions result = mockMvc.perform(
-                post("/notices")
+                post("/eventNotices")
                         .contentType(APPLICATION_JSON)
+                        .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN)
+                        .cookie(new Cookie("refresh-token", REFRESH_TOKEN))
                         .content(objectMapper.writeValueAsString(request))
         );
 
         // then
         result.andExpect(status().isCreated())
                 .andDo(restDocs.document(
+                        requestCookies(
+                                cookieWithName("refresh-token")
+                                        .description("갱신 토큰")
+                        ),
+                        requestHeaders(
+                                headerWithName("Authorization")
+                                        .description("access token")
+                        ),
                         requestFields(
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("공지 사항 내용"),
-                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
+                                fieldWithPath("title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content").type(JsonFieldType.STRING).description("이벤트 공지 사항 내용"),
+                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
                                 fieldWithPath("fileIds").type(JsonFieldType.ARRAY).description("첨부 파일 ID 목록").optional()
                         ),
                         responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("공지 사항 ID"),
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("공지 사항 내용"),
-                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("공지 사항 조회수"),
-                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
-                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("공지 사항 생성일"),
-                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("공지 사항 수정일"),
+                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 ID"),
+                                fieldWithPath("title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content").type(JsonFieldType.STRING).description("이벤트 공지 사항 내용"),
+                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 조회수"),
+                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
+                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 생성일"),
+                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 수정일"),
                                 fieldWithPath("files").type(JsonFieldType.ARRAY).description("첨부 파일 목록"),
                                 fieldWithPath("files[].id").type(JsonFieldType.NUMBER).description("파일 ID"),
                                 fieldWithPath("files[].uuid").type(JsonFieldType.STRING).description("파일 고유 식별자"),
@@ -98,20 +115,20 @@ class NoticeControllerTest extends AbstractControllerTest {
                 ));
     }
 
-    @DisplayName("공지 사항 리스트를 조회할 수 있다.")
+    @DisplayName("이벤트 공지 사항 리스트를 조회할 수 있다.")
     @Test
-    void getNoticeList() throws Exception {
+    void getEventNoticeList() throws Exception {
 
         // given
-        NoticeListElementResponse notice1 = new NoticeListElementResponse(1L, "공지 사항 1", 10, true, LocalDateTime.now(), LocalDateTime.now());
-        NoticeListElementResponse notice2 = new NoticeListElementResponse(2L, "공지 사항 2", 10, false, LocalDateTime.now(), LocalDateTime.now());
-        Page<NoticeListElementResponse> page = new PageImpl<>(List.of(notice1, notice2), PageRequest.of(0, 10), 2);
+        EventNoticeListElementResponse eventNotice1 = new EventNoticeListElementResponse(1L, "이벤트 공지 사항 1", 10, true, LocalDateTime.now(), LocalDateTime.now());
+        EventNoticeListElementResponse eventNotice2 = new EventNoticeListElementResponse(2L, "이벤트 공지 사항 2", 10, false, LocalDateTime.now(), LocalDateTime.now());
+        Page<EventNoticeListElementResponse> page = new PageImpl<>(List.of(eventNotice1, eventNotice2), PageRequest.of(0, 10), 2);
 
-        when(noticeService.getNoticeList(any(), any(Pageable.class))).thenReturn(page);
+        when(eventNoticeService.getEventNoticeList(any(), any())).thenReturn(page);
 
         // when
         ResultActions result = mockMvc.perform(
-                get("/notices")
+                get("/eventNotices")
                         .contentType(APPLICATION_JSON)
         );
 
@@ -119,17 +136,17 @@ class NoticeControllerTest extends AbstractControllerTest {
         result.andExpect(status().isOk())
                 .andDo(restDocs.document(
                         queryParameters(
-                                parameterWithName("title").description("찾고자 하는 공지 사항 제목").optional(),
+                                parameterWithName("title").description("찾고자 하는 이벤트 공지 사항 제목").optional(),
                                 parameterWithName("page").description("페이지 번호 [default: 0]").optional(),
                                 parameterWithName("size").description("페이지 크기 [default: 10]").optional()
                         ),
                         responseFields(
-                                fieldWithPath("content[].id").type(JsonFieldType.NUMBER).description("공지 사항 ID"),
-                                fieldWithPath("content[].title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content[].hitCount").type(JsonFieldType.NUMBER).description("공지 사항 조회수"),
-                                fieldWithPath("content[].fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
-                                fieldWithPath("content[].createdAt").type(JsonFieldType.STRING).description("공지 사항 생성일"),
-                                fieldWithPath("content[].updatedAt").type(JsonFieldType.STRING).description("공지 사항 수정일"),
+                                fieldWithPath("content[].id").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 ID"),
+                                fieldWithPath("content[].title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content[].hitCount").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 조회수"),
+                                fieldWithPath("content[].fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
+                                fieldWithPath("content[].createdAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 생성일"),
+                                fieldWithPath("content[].updatedAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 수정일"),
                                 fieldWithPath("pageable").type(JsonFieldType.OBJECT).description("페이지 정보"),
                                 fieldWithPath("pageable.pageNumber").type(JsonFieldType.NUMBER).description("현재 페이지 번호"),
                                 fieldWithPath("pageable.pageSize").type(JsonFieldType.NUMBER).description("페이지 당 요소 수"),
@@ -154,9 +171,9 @@ class NoticeControllerTest extends AbstractControllerTest {
                 ));
     }
 
-    @DisplayName("공지 사항을 조회할 수 있다.")
+    @DisplayName("이벤트 공지 사항을 조회할 수 있다.")
     @Test
-    void getNotice() throws Exception {
+    void getEventNotice() throws Exception {
 
         // given
         List<FileResponse> files = Arrays.asList(
@@ -164,13 +181,13 @@ class NoticeControllerTest extends AbstractControllerTest {
                 new FileResponse(2L, "11a480c0-13fa-11ef-9047-570191b390ea", "예시 첨부 파일 2.jpg", "image/jpeg", LocalDateTime.now(), LocalDateTime.now()),
                 new FileResponse(3L, "1883fc70-cfb4-11ee-a387-e754bd392d45", "예시 첨부 파일 3.jpg", "image/jpeg", LocalDateTime.now(), LocalDateTime.now())
         );
-        NoticeResponse response = new NoticeResponse(1L, "공지 사항 제목", "content", 10, true, LocalDateTime.now(), LocalDateTime.now(), files);
+        EventNoticeResponse response = new EventNoticeResponse(1L, "이벤트 공지 사항 제목", "content", 10, true, LocalDateTime.now(), LocalDateTime.now(), files);
 
-        when(noticeService.getNotice(anyLong())).thenReturn(response);
+        when(eventNoticeService.getEventNotice(any())).thenReturn(response);
 
         // when
         ResultActions result = mockMvc.perform(
-                get("/notices/{noticeId}", 1L)
+                get("/eventNotices/{eventNoticeId}", 1L)
                         .contentType(APPLICATION_JSON)
         );
 
@@ -178,16 +195,16 @@ class NoticeControllerTest extends AbstractControllerTest {
         result.andExpect(status().isOk())
                 .andDo(restDocs.document(
                         pathParameters(
-                                parameterWithName("noticeId").description("조회할 공지 사항 ID")
+                                parameterWithName("eventNoticeId").description("조회할 이벤트 공지 사항 ID")
                         ),
                         responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("공지 사항 ID"),
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("공지 사항 내용"),
-                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("공지 사항 조회수"),
-                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
-                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("공지 사항 생성일"),
-                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("공지 사항 수정일"),
+                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 ID"),
+                                fieldWithPath("title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content").type(JsonFieldType.STRING).description("이벤트 공지 사항 내용"),
+                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 조회수"),
+                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
+                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 생성일"),
+                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 수정일"),
                                 fieldWithPath("files").type(JsonFieldType.ARRAY).description("첨부 파일 목록"),
                                 fieldWithPath("files[].id").type(JsonFieldType.NUMBER).description("파일 ID"),
                                 fieldWithPath("files[].uuid").type(JsonFieldType.STRING).description("파일 고유 식별자"),
@@ -200,48 +217,58 @@ class NoticeControllerTest extends AbstractControllerTest {
                 ));
     }
 
-    @DisplayName("공지 사항을 수정할 수 있다.")
+    @DisplayName("이벤트 공지 사항을 수정할 수 있다.")
     @Test
-    void updateNotice() throws Exception {
+    void updateEventNotice() throws Exception {
 
         // given
-        NoticeRequest request = new NoticeRequest("수정된 공지 사항 제목", "수정된 공지 사항 내용", false, List.of(1L, 2L, 3L));
+        EventNoticeRequest request = new EventNoticeRequest("수정된 이벤트 공지 사항 제목", "수정된 이벤트 공지 사항 내용", false, List.of(1L, 2L, 3L));
         List<FileResponse> files = Arrays.asList(
                 new FileResponse(1L, "014eb8a0-d4a6-11ee-adac-117d766aca1d", "예시 첨부 파일 1.jpg", "image/jpeg", LocalDateTime.of(2024, 1, 1, 12, 0), LocalDateTime.now()),
                 new FileResponse(2L, "11a480c0-13fa-11ef-9047-570191b390ea", "예시 첨부 파일 2.jpg", "image/jpeg", LocalDateTime.of(2024, 1, 1, 12, 0), LocalDateTime.now()),
                 new FileResponse(3L, "1883fc70-cfb4-11ee-a387-e754bd392d45", "예시 첨부 파일 3.jpg", "image/jpeg", LocalDateTime.of(2024, 1, 1, 12, 0), LocalDateTime.now())
         );
-        NoticeResponse response = new NoticeResponse(1L, "수정된 공지 사항 제목", "수정된 공지 사항 내용", 10, false, LocalDateTime.of(2024, 1, 1, 12, 0), LocalDateTime.now(), files);
+        EventNoticeResponse response = new EventNoticeResponse(1L, "수정된 이벤트 공지 사항 제목", "수정된 이벤트 공지 사항 내용", 10, false, LocalDateTime.of(2024, 1, 1, 12, 0), LocalDateTime.now(), files);
 
-        when(noticeService.updateNotice(anyLong(), any(NoticeRequest.class))).thenReturn(response);
+        when(eventNoticeService.updateEventNotice(anyLong(), any(EventNoticeRequest.class))).thenReturn(response);
 
         // when
         ResultActions result = mockMvc.perform(
-                put("/notices/{noticeId}", 1L)
+                put("/eventNotices/{eventNoticeId}", 1L)
                         .contentType(APPLICATION_JSON)
+                        .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN)
+                        .cookie(new Cookie("refresh-token", REFRESH_TOKEN))
                         .content(objectMapper.writeValueAsString(request))
         );
 
         // then
         result.andExpect(status().isOk())
                 .andDo(restDocs.document(
+                        requestCookies(
+                                cookieWithName("refresh-token")
+                                        .description("갱신 토큰")
+                        ),
+                        requestHeaders(
+                                headerWithName("Authorization")
+                                        .description("access token")
+                        ),
                         pathParameters(
-                                parameterWithName("noticeId").description("수정할 공지 사항 ID")
+                                parameterWithName("eventNoticeId").description("수정할 이벤트 공지 사항 ID")
                         ),
                         requestFields(
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("공지 사항 내용"),
-                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
+                                fieldWithPath("title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content").type(JsonFieldType.STRING).description("이벤트 공지 사항 내용"),
+                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
                                 fieldWithPath("fileIds").type(JsonFieldType.ARRAY).description("첨부 파일 ID 목록").optional()
                         ),
                         responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("공지 사항 ID"),
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("공지 사항 제목"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("공지 사항 내용"),
-                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("공지 사항 조회수"),
-                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("공지 사항 고정 여부"),
-                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("공지 사항 생성일"),
-                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("공지 사항 수정일"),
+                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 ID"),
+                                fieldWithPath("title").type(JsonFieldType.STRING).description("이벤트 공지 사항 제목"),
+                                fieldWithPath("content").type(JsonFieldType.STRING).description("이벤트 공지 사항 내용"),
+                                fieldWithPath("hitCount").type(JsonFieldType.NUMBER).description("이벤트 공지 사항 조회수"),
+                                fieldWithPath("fixed").type(JsonFieldType.BOOLEAN).description("이벤트 공지 사항 고정 여부"),
+                                fieldWithPath("createdAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 생성일"),
+                                fieldWithPath("updatedAt").type(JsonFieldType.STRING).description("이벤트 공지 사항 수정일"),
                                 fieldWithPath("files").type(JsonFieldType.ARRAY).description("첨부 파일 목록"),
                                 fieldWithPath("files[].id").type(JsonFieldType.NUMBER).description("파일 ID"),
                                 fieldWithPath("files[].uuid").type(JsonFieldType.STRING).description("파일 고유 식별자"),
@@ -255,24 +282,34 @@ class NoticeControllerTest extends AbstractControllerTest {
     }
 
 
-    @DisplayName("공지 사항을 삭제할 수 있다.")
+    @DisplayName("이벤트 공지 사항을 삭제할 수 있다.")
     @Test
-    void deleteNotice() throws Exception {
+    void deleteEventNotice() throws Exception {
 
         // given
-        doNothing().when(noticeService).deleteNotice(anyLong());
+        doNothing().when(eventNoticeService).deleteEventNotice(anyLong());
 
         // when
         ResultActions result = mockMvc.perform(
-                delete("/notices/{noticeId}", 1L)
+                delete("/eventNotices/{eventNoticeId}", 1L)
                         .contentType(APPLICATION_JSON)
+                        .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN)
+                        .cookie(new Cookie("refresh-token", REFRESH_TOKEN))
         );
 
         // then
         result.andExpect(status().isNoContent())
                 .andDo(restDocs.document(
+                        requestCookies(
+                                cookieWithName("refresh-token")
+                                        .description("갱신 토큰")
+                        ),
+                        requestHeaders(
+                                headerWithName("Authorization")
+                                        .description("access token")
+                        ),
                         pathParameters(
-                                parameterWithName("noticeId").description("삭제할 공지 사항 ID")
+                                parameterWithName("eventNoticeId").description("삭제할 이벤트 공지 사항 ID")
                         )
                 ));
     }
